cosaves: Successfully load an entire cosave
As far as I can tell, this is the first time Wrye Bash has actually
loaded an _entire_ cosave, including all plugins. Previously, it just
aborted once it finished reading the xSE section. This opens the way to
changing more than just the MODS entry in the xSE section - I've already
identified several other plugins for Skyrim that store plugin names as
well that should be remapped.

Of course, it also opens the way for a significant performance
improvement - we could keep a list of only the relevant chunks around
and then quickly remap when needed.